+++ src/index.js
@@ -163,7 +163,7 @@ function deleteOrganisationDevice(
   let urlParts = ['organisations', organisationId, 'devices', deviceId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -226,7 +226,7 @@ function getOrganisationDevice(
   let urlParts = ['organisations', organisationId, 'devices', deviceId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -298,7 +298,7 @@ function putOrganisationDevice(
   let urlParts = ['organisations', organisationId, 'devices', deviceId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -383,7 +383,7 @@ function deleteOrganisationNetworkDeviceAccess(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -468,7 +468,7 @@ function putOrganisationNetworkDeviceAccess(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -529,7 +529,7 @@ function deleteUserNotificationRead(
   let urlParts = ['users', userId, 'notifications', notificationId, 'read'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -590,7 +590,7 @@ function putUserNotificationRead(
   let urlParts = ['users', userId, 'notifications', notificationId, 'read'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -673,7 +673,7 @@ function deleteUserOrganisationAggregation(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -759,7 +759,7 @@ function putUserOrganisationAggregation(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -840,7 +840,7 @@ function deleteUserOrganisationAlert(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -924,7 +924,7 @@ function putUserOrganisationAlert(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1018,7 +1018,7 @@ function deleteUserOrganisationDeviceModule(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1115,7 +1115,7 @@ function putUserOrganisationDeviceModule(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1198,7 +1198,7 @@ function deleteUserOrganisationInvitation(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1284,7 +1284,7 @@ function putUserOrganisationInvitation(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1367,7 +1367,7 @@ function deleteUserOrganisationMember(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1453,7 +1453,7 @@ function putUserOrganisationMember(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1536,7 +1536,7 @@ function putUserOrganisationOpenNetworkMembership(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1617,7 +1617,7 @@ function deleteUserOrganisationShare(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1701,7 +1701,7 @@ function putUserOrganisationShare(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1750,7 +1750,7 @@ function getOpenAPI(
   let urlParts = ['openAPI'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1791,7 +1791,7 @@ function getCrops({ xAppVersion } = {}, options) {
   let urlParts = ['crops'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -1828,7 +1828,7 @@ function getModules({ xAppVersion } = {}, options) {
   let urlParts = ['modules'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -1869,7 +1869,7 @@ function getMySelf({ authorization, accessToken, xAppVersion } = {}, options) {
   let urlParts = ['me'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -1917,7 +1917,7 @@ function getNetwork({ networkId, xAppVersion } = {}, options) {
   let urlParts = ['networks', networkId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -1984,7 +1984,7 @@ function getOAuth2Authorize(
   let urlParts = ['oauth2', 'authorize'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({
@@ -2042,7 +2042,7 @@ function getOrganisation(
   let urlParts = ['organisations', organisationId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2099,7 +2099,7 @@ function putOrganisation(
   let urlParts = ['organisations', organisationId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2159,7 +2159,7 @@ function getOrganisationChargebeePortalSession(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2233,7 +2233,7 @@ function getOrganisationDevices(
   let urlParts = ['organisations', organisationId, 'devices'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2339,7 +2339,7 @@ function getOrganisationDevicesLiveAggregations(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2416,7 +2416,7 @@ function getOrganisationHostedPage(
   let urlParts = ['organisations', organisationId, 'hostedPage'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2487,7 +2487,7 @@ function getOrganisationInvitation(
   let urlParts = ['organisations', organisationId, 'invitations', invitationId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2567,7 +2567,7 @@ function getOrganisationInvitations(
   let urlParts = ['organisations', organisationId, 'invitations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2634,7 +2634,7 @@ function getOrganisationMember(
   let urlParts = ['organisations', organisationId, 'members', memberId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2689,7 +2689,7 @@ function getOrganisationMembers(
   let urlParts = ['organisations', organisationId, 'members'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2781,7 +2781,7 @@ function getOrganisationNetworkDevices(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2858,7 +2858,7 @@ function getOrganisationNetworks(
   let urlParts = ['organisations', organisationId, 'networks'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2921,7 +2921,7 @@ function getOrganisationPlace(
   let urlParts = ['organisations', organisationId, 'place', placeId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -2976,7 +2976,7 @@ function getOrganisationPlaces(
   let urlParts = ['organisations', organisationId, 'places'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3033,7 +3033,7 @@ function postOrganisationPlace(
   let urlParts = ['organisations', organisationId, 'places'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3088,7 +3088,7 @@ function getOrganisationPreferences(
   let urlParts = ['organisations', organisationId, 'preferences'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3145,7 +3145,7 @@ function putOrganisationPreferences(
   let urlParts = ['organisations', organisationId, 'preferences'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3206,7 +3206,7 @@ function getOrganisationShare(
   let urlParts = ['organisations', organisationId, 'shares', shareId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3289,7 +3289,7 @@ function getOrganisationShares(
   let urlParts = ['organisations', organisationId, 'shares'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3361,7 +3361,7 @@ function getPartnerDevices(
   let urlParts = ['partners', partnerId, 'devices'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3403,7 +3403,7 @@ function getPing({ xAppVersion } = {}, options) {
   let urlParts = ['ping'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -3520,7 +3520,7 @@ function getPublicWeatherLive(
   let urlParts = ['weather', 'public', 'live'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3574,7 +3574,7 @@ function getQuote({ quoteId, xAppVersion } = {}, options) {
   let urlParts = ['quotes', quoteId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -3642,7 +3642,7 @@ function getQuoteHostedPage(
   let urlParts = ['users', userId, 'quotes', quoteId, 'hostedPage'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3700,7 +3700,7 @@ function getSearchDevices({ latitude, longitude, xAppVersion } = {}, options) {
   let urlParts = ['search', 'devices'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({
@@ -3753,7 +3753,7 @@ function getSearchUser(
   let urlParts = ['search', 'users'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3807,7 +3807,7 @@ function getUser(
   let urlParts = ['users', userId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3862,7 +3862,7 @@ function putUser(
   let urlParts = ['users', userId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3932,7 +3932,7 @@ function getUserAggregation(
   let urlParts = ['users', userId, 'aggregations', aggregationId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -3990,7 +3990,7 @@ function getUserAggregations(
   let urlParts = ['users', userId, 'aggregations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4045,7 +4045,7 @@ function getUserAlerts(
   let urlParts = ['users', userId, 'alerts'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4161,7 +4161,7 @@ function getUserDailyData(
   let urlParts = ['users', userId, 'data', 'daily'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4241,7 +4241,7 @@ function getUserDevice(
   let urlParts = ['users', userId, 'devices', deviceId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4355,7 +4355,7 @@ function getUserDeviceContinuousStatistics(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4467,7 +4467,7 @@ function getUserDeviceDailyData(
   let urlParts = ['users', userId, 'devices', deviceId, 'data', 'daily'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4541,7 +4541,7 @@ function getUserDeviceForecasts(
   let urlParts = ['users', userId, 'devices', deviceId, 'forecasts'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4646,7 +4646,7 @@ function getUserDeviceHourlyData(
   let urlParts = ['users', userId, 'devices', deviceId, 'data', 'hourly'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4714,7 +4714,7 @@ function getUserDeviceModules(
   let urlParts = ['users', userId, 'devices', deviceId, 'modules'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4784,7 +4784,7 @@ function getUserDevicePositions(
   let urlParts = ['users', userId, 'devices', deviceId, 'positions'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4880,7 +4880,7 @@ function getUserDeviceRawData(
   let urlParts = ['users', userId, 'devices', deviceId, 'data', 'raw'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -4991,7 +4991,7 @@ function getUserDeviceStatistics(
   let urlParts = ['users', userId, 'devices', deviceId, 'statistics'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5085,7 +5085,7 @@ function getUserDeviceSummary(
   let urlParts = ['users', userId, 'devices', deviceId, 'summaries', date];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5163,7 +5163,7 @@ function getUserDevices(
   let urlParts = ['users', userId, 'devices'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5266,7 +5266,7 @@ function getUserDevicesLiveAggregations(
   let urlParts = ['users', userId, 'devices', 'liveAggregations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5352,7 +5352,7 @@ function getUserForecasts(
   let urlParts = ['users', userId, 'forecasts'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5471,7 +5471,7 @@ function getUserHourlyData(
   let urlParts = ['users', userId, 'data', 'hourly'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5542,7 +5542,7 @@ function getUserInvitation(
   let urlParts = ['users', userId, 'invitations', invitationId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5623,7 +5623,7 @@ function getUserInvitations(
   let urlParts = ['users', userId, 'invitations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5685,7 +5685,7 @@ function getUserNotifications(
   let urlParts = ['users', userId, 'notifications'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5740,7 +5740,7 @@ function getUserOrganisations(
   let urlParts = ['users', userId, 'organisations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5795,7 +5795,7 @@ function postUserOrganisation(
   let urlParts = ['users', userId, 'organisations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5856,7 +5856,7 @@ function getUserPendingInvitation(
   let urlParts = ['users', userId, 'pendingInvitations', invitationId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5921,7 +5921,7 @@ function getUserPendingInvitations(
   let urlParts = ['users', userId, 'pendingInvitations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -5976,7 +5976,7 @@ function getUserPreferences(
   let urlParts = ['users', userId, 'preferences'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6031,7 +6031,7 @@ function putUserPreferences(
   let urlParts = ['users', userId, 'preferences'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6154,7 +6154,7 @@ function getUserStatistics(
   let urlParts = ['users', userId, 'statistics'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6235,7 +6235,7 @@ function getWeatherLive(
   let urlParts = ['weather', 'live'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6280,7 +6280,7 @@ function postLogin({ body, xAppVersion } = {}, options) {
   let urlParts = ['users', 'sign_in'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -6319,7 +6319,7 @@ function postLostPassword({ body, xAppVersion } = {}, options) {
   let urlParts = ['password', 'lost'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -6361,7 +6361,7 @@ function postOAuth2Token({ body, authorization, xAppVersion } = {}, options) {
   let urlParts = ['oauth2', 'token'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6408,7 +6408,7 @@ function postOrganisation(
   let urlParts = ['organisations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6465,7 +6465,7 @@ function postOrganisationPaymentComplete(
   let urlParts = ['organisations', organisationId, 'paymentComplete'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6520,7 +6520,7 @@ function postPartnerToken(
   let urlParts = ['partners', partnerId, 'tokens'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6573,7 +6573,7 @@ function postPartnerTokenRequest(
   let urlParts = ['partners', partnerId, 'tokenRequests'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6630,7 +6630,7 @@ function postQuoteOffline(
   let urlParts = ['users', userId, 'quotes', quoteId, 'offline'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6679,7 +6679,7 @@ function postSignatureCloudinary(
   let urlParts = ['cloudinary', 'signature'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6723,7 +6723,7 @@ function postSignupCheck({ body, authorization, xAppVersion } = {}, options) {
   let urlParts = ['utils', 'checkSignup'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6770,7 +6770,7 @@ function postSocialImage(
   let urlParts = ['socialImage'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6812,7 +6812,7 @@ function postUser({ body, xAppVersion } = {}, options) {
   let urlParts = ['users'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -6882,7 +6882,7 @@ function postUserDeviceExport(
   let urlParts = ['users', userId, 'devices', deviceId, 'export'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -6942,7 +6942,7 @@ function postUserNetworkInvitation(
   let urlParts = ['users', userId, 'networks', networkId, 'invitations'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -7015,7 +7015,7 @@ function postUserOrganisationAggregation(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7085,7 +7085,7 @@ function postUserOrganisationAlert(
   let urlParts = ['users', userId, 'organisations', organisationId, 'alerts'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7155,7 +7155,7 @@ function postUserOrganisationDevice(
   let urlParts = ['users', userId, 'organisations', organisationId, 'devices'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7231,7 +7231,7 @@ function postUserOrganisationInvitation(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7301,7 +7301,7 @@ function postUserOrganisationMember(
   let urlParts = ['users', userId, 'organisations', organisationId, 'members'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7349,7 +7349,7 @@ function postUserPassword({ body, userId, xAppVersion } = {}, options) {
   let urlParts = ['users', userId, 'password'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -7388,7 +7388,7 @@ function postVerify({ body, xAppVersion } = {}, options) {
   let urlParts = ['users', 'verify'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -7478,7 +7478,7 @@ function putOrganisationDeviceCalibration(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7557,7 +7557,7 @@ function putOrganisationDeviceSettings(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7627,7 +7627,7 @@ function putOrganisationPlace(
   let urlParts = ['organisations', organisationId, 'places', placeId];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7727,7 +7727,7 @@ function putPartnerModuleParameters(
   ];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7769,7 +7769,7 @@ function putPassword({ body, xAppVersion } = {}, options) {
   let urlParts = ['password'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     'X-APP-Version': xAppVersion,
   });
   let qs = cleanQuery({});
@@ -7829,7 +7829,7 @@ function putUserInvitationAnswer(
   let urlParts = ['users', userId, 'invitations', invitationId, 'answer'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7893,7 +7893,7 @@ function getGraphQL(
   let urlParts = ['graphql'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });
@@ -7945,7 +7945,7 @@ function postGraphQL(
   let urlParts = ['graphql'];
   let headers = Object.assign((options || {}).headers || {}, {
     'X-API-Version': '1.49.6',
-    'X-SDK-Version': '2.4.0',
+    'X-SDK-Version': '2.5.0',
     Authorization: authorization,
     'X-APP-Version': xAppVersion,
   });